feat: drag ruler integration (#984)
This release includes built-in support for the module "Drag Ruler" (https://foundryvtt.com/packages/drag-ruler).

For travellers, each of the supported rulesets has a behavior for default movement. An optional setting (under display settings) allows for encumbrance calculations to affect the movement range displayed through the module. Also, default movement units and values are now part of the rules settings and are used when new actors are created.  For example, the default value is 6m for a minor action.  Other rulesets use movement for an action (e.g. 10m for CD).

For ships, movement ruler will use the Jump value on the ship sheet IF the scenes's grid scale units is set to "pc" (for parsecs).  This is useful for jumps on subsector maps.