feat: basic skills select (#1300)
This update improves the roll dialog.  Rather than only being able to edit the skill level (value), the user can instead select a different skill from a list of valid skills from that actor prior to the roll. The roll dialog will now show the skill chosen. This is useful where the default skill is not correct, missing, or should be changed based on circumstances.  Thanks to the suggestion of @romanK