feat:  Add Target DM roll setting (#1512)
This release adds a new feature where GM's can specify a list of custom modifiers (as new string setting) to be used for item rolls.  Such modifiers are typically Target DM (e.g. target is prone).  For the open rules sets, some of the more common modifiers have been pre-populated.  

The modifier list appears as a select list in the roll dialog.  The details of this new feature are found here:
https://github.com/xdy/twodsix-foundryvtt/wiki/Custom-Target-Modifiers
Thanks to @Zee for the suggestion.

This release also includes Spanish localization updates courtesy of @ForjaSalvaje