feat: Make Traveller and Ship Sheet Responsive (#1566)
This update (the final update before the v12 baseline change) allows the Traveller and Ship sheet to be resizable.  Please note that there will be some minor visual changes and the scaling doesn't alter the font size.  As always, please try out this version before running a game.  Amy issues/bugs please make certain to show a screen capture.  

This update also allows powerUsed field to be a decimal (per suggestion of @Finderski)